Migrate bot_instance
